# docs/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Factory Ops API
  description: |
    工廠值班工作管理系統 REST API(v1.3)。
    - API-first,Web 與未來 Native App 共用同一組 API。
    - 認證採 JWT Bearer Token(access + refresh)。JWT claims 帶 `userId`、`accountName`、`rootOrgId`、`orgPath[]`、`roles[]`、`groupIds[]`、`orgManagerScopes[]`(衍生自 Organization.managerId 反查)。
    - **多租戶隔離**:所有資源(除 `Organization` 自身)歸屬於某個 `rootOrgId`,server 強制 scope。跨 root org 操作僅 `ADMIN` 系統角色可進行。
    - **時間欄位**:wire format 使用 ISO 8601 + offset(例 `2026-05-04T08:30:00+08:00`,Q-17 拍板);儲存層為 UTC `Instant`。
    - 列表端點預設使用 cursor-based pagination,支援欄位投影與增量同步。
    - **Organization 為 polymorphic 樹**(FAB → DIVISION → DEPARTMENT → SECTION),只有 leaf type 可以承載 Group / Project / Task(見 ADR-0004)。每節點有 **單一 `managerId`**(對應 ORG_MANAGER)+ **多位 `leaderIds[]`**(業務負責人列表)— 見 ADR-0010。
    - **Group 為平面結構**(屬唯一 leaf Organization),polymorphic by `type`(DEFAULT / LINE / TEAM / SHIFT…);Group 加 `settings.qa` 用於設定 QA 雙簽機制(見 ADR-0011)。
    - **Template 機制**:支援 GLOBAL(ADMIN 維護)與 ORG(各 Org 維護)兩種 scope;可 fork from GLOBAL,可直接實例化(見 ADR-0006)。
    - **跨層派工(v1.3 改寫)**:任一 Organization 節點的 manager 可直接派 ActionRequest 到其子孫中的 leaf Org(`targetOrgId` 必為 leaf);**v1.3 移除多層 relay 流程**(見 ADR-0008 v1.3)。
    - **Task QA 雙簽**:Task 建立時從所屬 Project.groupIds 對應 Group settings.qa snapshot policy;之後 Group settings 變更不影響該 Task;Review action 走 `POST /tasks/{taskId}/review`(見 ADR-0011)。
    - **HR 整合**:User 為 HR User 的本地投影,以 `accountName` 對 HR Mock REST 查詢(見 ADR-0007 v1.3 Amendment)。
...
  version: 1.3.0
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: 以 accountName + 密碼登入
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登入成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 以 refresh token 換發新的 access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: 新 token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: 登出(撤銷 refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '204': { description: No Content }
  /auth/password:
    put:
      tags: [Auth]
      summary: 變更自己的密碼
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password, minLength: 12 }
      responses:
        '204': { description: No Content }
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ===================== Me =====================
...
  /orgs/{orgId}/dispatch-action-request:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    post:
      tags: [ActionRequests]
      summary: 上級向下級 leaf 直接派 ActionRequest(Single-hop;v1.3 改寫)
      description: |
        - `orgId` 是 **targetOrgId**(被派的 leaf Organization)。
        - **target 必為 leaf**(`type ∈ root.settings.leafTypes`),否則 422 `target_must_be_leaf`。
        - actor 必須是 target 某個祖先節點(含 target 本身)的 manager(`Organization.managerId == actor.userId`),否則 403 `not_authorized_to_dispatch`。server 自動推導 `originatingOrgId` 為「actor 所有 manager scope 中,target 的最近祖先」。
        - **預設 ownerId 規則**(見 ADR-0010):
          - `targetOrg.leaderIds.length == 0` → 409 `target_org_no_leader`
          - `length == 1` → 自動 `ownerId = leaderIds[0]`(忽略 body.ownerId 或要求等於該值)
          - `length >= 2` → 必須 body.ownerId ∈ leaderIds;未提供則 422 `owner_must_be_specified`,不在集合中則 422 `owner_not_in_leaders`
        - **v1.3 移除「relay 到下一級」流程**;不再有 `POST /action-requests/{id}/relay`。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DispatchActionRequestRequest' }
      responses:
        '201':
          description: 已建立 ActionRequest
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRequest' }
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ===================== Groups =====================
...
  /action-requests:
    get:
      tags: [ActionRequests]
      summary: 列出 ActionRequest
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/FieldsParam'
        - in: query
          name: projectId
          schema: { type: string }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/ActionRequestStatus' }
        - in: query
          name: severity
          schema: { $ref: '#/components/schemas/SeverityLevel' }
        - in: query
          name: requesterId
          schema: { type: string }
      responses:
        '200':
          description: ActionRequest 清單
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRequestPage' }
    post:
      tags: [ActionRequests]
      summary: 提交 ActionRequest
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateActionRequestRequest' }
      responses:
        '201':
          description: 已提交
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRequest' }
  /action-requests/{actionRequestId}:
    parameters:
      - in: path
        name: actionRequestId
        required: true
        schema: { type: string }
    get:
      tags: [ActionRequests]
      summary: 讀取單一 ActionRequest
      responses:
        '200':
          description: ActionRequest
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRequest' }
    patch:
      tags: [ActionRequests]
      summary: 更新內容
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateActionRequestRequest' }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRequest' }
    delete:
      tags: [ActionRequests]
      summary: 軟刪除
      responses:
        '204': { description: No Content }
  /action-requests/{actionRequestId}/status:
    parameters:
      - in: path
        name: actionRequestId
        required: true
        schema: { type: string }
    post:
      tags: [ActionRequests]
      summary: 變更狀態
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/ActionRequestStatus' }
                reason: { type: string }
      responses:
        '200':
          description: 已變更
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionRequest' }

  # /action-requests/{actionRequestId}/relay
  # ★ v1.3 已移除(Q-14 拍板:targetOrgId 必為 leaf,改為 single-hop direct dispatch;見 ADR-0008 v1.3)
  /action-requests/{actionRequestId}/convert-to-task:
    parameters:
      - in: path
        name: actionRequestId
        required: true
        schema: { type: string }
    post:
      tags: [ActionRequests]
      summary: 將 ActionRequest 轉換為 Task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConvertActionRequestToTaskRequest' }
      responses:
        '201':
          description: 已轉換,回傳新 Task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '409':
          $ref: '#/components/responses/Conflict'

  # ===================== Attachments =====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrgIdParam:
      in: path
      name: orgId
      required: true
      schema: { type: string }
      description: |
        Organization ID。在 `/orgs/{orgId}/...` 子路徑中,根據端點語意可能是 root org 或 任一節點:
        - `/orgs/{orgId}/groups`、`/orgs/{orgId}/project-templates` 等管理面 API:`orgId` 為 **rootOrgId**
        - `/orgs/{orgId}/dispatch-action-request`、`/orgs/{orgId}/transfer-leader`:`orgId` 為**特定節點**(可能任一層)
...
    CursorParam:
      in: query
      name: cursor
      schema: { type: string }
      description: cursor-based pagination 游標
    LimitParam:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    FieldsParam:
      in: query
      name: fields
      schema: { type: string }
      description: 逗號分隔的欄位投影,例 `id,title,status,ownerId`
...
  responses:
    Unauthorized:
      description: 未認證
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: 無權限
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: 業務規則衝突(狀態流、唯一性)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: 輸入驗證失敗
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
  schemas:
...
    Problem:
      description: RFC 7807 problem details
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
      required: [title, status]
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  code: { type: string }
                  message: { type: string }
    PageInfo:
      type: object
      properties:
        nextCursor: { type: string, nullable: true }
        hasMore: { type: boolean }
...
    AttachmentRef:
      type: object
      required: [attachmentId]
      properties:
        attachmentId: { type: string }
        alt: { type: string }
        role:
          type: string
          enum: [INLINE, ATTACHMENT, COVER]
          default: ATTACHMENT
    TimeRange:
      type: object
      properties:
        start: { type: string, format: date-time, nullable: true }
        due: { type: string, format: date-time, nullable: true }

    # ---------- Auth ----------
    LoginRequest:
      type: object
      required: [accountName, password]
      properties:
        accountName: { type: string, maxLength: 30, description: HR 識別字 }
        password: { type: string, format: password }
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresIn, tokenType]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: access token 秒數 }
        tokenType: { type: string, example: Bearer }

    # ---------- Enums ----------
    Role:
      type: string
      enum: [OPERATOR, SHIFT_LEAD, ENGINEER, QA, GROUP_ADMIN, GROUP_MANAGER, ORG_MANAGER, ORG_ADMIN, ADMIN]
      description: |
        - `OPERATOR` / `SHIFT_LEAD` / `ENGINEER` / `QA`:第一線執行
        - `GROUP_ADMIN`:Group 內 Template 與成員角色設定;Group settings(含 QA 雙簽)維護
        - `GROUP_MANAGER`:Group 經理,可指派 GROUP_ADMIN 代理;Group settings 最終設定權;權限 ⊇ GROUP_ADMIN
        - `ORG_MANAGER`:某 Organization 節點的唯一經理(衍生自 `Organization.managerId == userId`);可直接派 ActionRequest 到其子孫中的 leaf Org
        - `ORG_ADMIN`:該 root org 的最高權限;管理 Org 樹 / users / managers / leaders
        - `ADMIN`:系統管理員(SaaS 維運),唯一可維護 GLOBAL Templates
        > v1.3 注意:`ORG_MANAGER` 為衍生角色,不應由前端直接寫入 `User.roles[]`;由 server 從 `Organization.managerId` 反查並維護 `User.orgManagerScopes[]` 衍生欄位。
...
    TemplateScope:
      type: string
      enum: [GLOBAL, ORG]
      description: GLOBAL 由 ADMIN 維護,所有 Org 可見;ORG 屬某 root org
...
    TaskStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, BLOCKED, IN_REVIEW, DONE, CANCELLED]
    Priority:
      type: string
      enum: [LOW, NORMAL, HIGH, URGENT]
    ActionRequestStatus:
      type: string
      enum: [SUBMITTED, TRIAGED, IN_PROGRESS, RESOLVED, REJECTED]
    SeverityLevel:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]

    # ---------- Organization ----------
...
    TemplateRef:
      type: object
      description: 由 Template 實例化時記錄的 snapshot,稽核用,不形成執行期相依
      properties:
        templateType:
          type: string
          enum: [PROJECT_TEMPLATE, TASK_TEMPLATE]
        templateScope: { $ref: '#/components/schemas/TemplateScope' }
        templateId: { type: string }
        templateVersion: { type: integer }
        instantiatedAt: { type: string, format: date-time }

    # ---------- Task ----------
    Task:
      type: object
      properties:
        id: { type: string }
        rootOrgId: { type: string }
        projectId: { type: string }
        type:
          type: string
          description: 多型 discriminator,例 EQUIPMENT_INSPECTION
        attributes:
          type: object
          additionalProperties: true
          description: type-specific 屬性,以對應 JSON Schema 驗證
        title: { type: string, description: 允許 Unicode 多語系 }
        descriptionMarkdown: { type: string }
        status: { $ref: '#/components/schemas/TaskStatus' }
        priority: { $ref: '#/components/schemas/Priority' }
        ownerId: { type: string }
        assignees:
          type: array
          minItems: 1
          items: { type: string }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/AttachmentRef' }
        schedule: { $ref: '#/components/schemas/TimeRange' }
        completedAt: { type: string, format: date-time, nullable: true }
        originActionRequestId: { type: string, nullable: true }
        templateRef:
          allOf:
            - $ref: '#/components/schemas/TemplateRef'
          nullable: true
        qaReviewPolicy:
          $ref: '#/components/schemas/QaReviewPolicy'
        qaReviews:
          type: array
          description: append-only QA 雙簽簽核紀錄(v1.3 新增)
          items: { $ref: '#/components/schemas/QaReviewEntry' }
        tags:
          type: array
          items: { type: string }
        overdue: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
        schemaVersion: { type: integer }
      required: [id, rootOrgId, projectId, type, title, status, priority, ownerId, assignees, qaReviewPolicy]
    QaReviewPolicy:
      type: object
      description: |
        Task 建立時(或 ActionRequest convert 時)從所屬 Project.groupIds[].settings.qa 合併 snapshot 的政策。
        後續 Group settings 變更**不影響**本 Task(對應 INV-31)。
        多 Group 合併規則:dualSignRequired = OR(各 group);requiredReviewerRoles = 各 group 的聯集去重。
      required: [dualSignRequired, requiredReviewerRoles, snapshotAt]
      properties:
        dualSignRequired: { type: boolean }
        requiredReviewerRoles:
          type: array
          items: { $ref: '#/components/schemas/Role' }
        snapshotAt: { type: string, format: date-time, description: snapshot 時間(ISO 8601 + offset) }
        sourceGroupIds:
          type: array
          description: 來源 Group ID 清單(稽核用)
          items: { type: string }
    QaReviewEntry:
      type: object
      description: 一筆 QA review action 記錄
      required: [reviewerId, reviewerRole, decision, at]
      properties:
        reviewerId: { type: string }
        reviewerRole: { $ref: '#/components/schemas/Role' }
        decision: { $ref: '#/components/schemas/ReviewDecision' }
        reason: { type: string }
        at: { type: string, format: date-time }
    ReviewDecision:
      type: string
      enum: [APPROVED, REJECTED]
...
    ActionRequest:
      type: object
      properties:
        id: { type: string }
        rootOrgId: { type: string }
        projectId:
          type: string
          nullable: true
          description: 自課內提報時必填;跨層 dispatch 起初可空,leaf triage 時關聯 Project
        originatingOrgId:
          type: string
          description: |
            發起者所掛的 Organization 節點(actor 是該節點的 manager,且該節點是 targetOrg 的祖先;
            自課內提報時 = targetOrgId)
        targetOrgId:
          type: string
          description: |
            **v1.3 改名**(取代 v1.2 `assignedToOrgId`)。**必為 leaf Organization**(`type ∈ root.settings.leafTypes`)。
            一旦寫入後不再變動(無中途 relay 流程)。
        title: { type: string, description: 允許 Unicode 多語系 }
        descriptionMarkdown: { type: string }
        severity:
          type: object
          properties:
            level: { $ref: '#/components/schemas/SeverityLevel' }
            reason: { type: string }
        status: { $ref: '#/components/schemas/ActionRequestStatus' }
        requesterId: { type: string }
        ownerId:
          type: string
          nullable: true
          description: |
            預設規則(對應 ADR-0010):
            - targetOrg.leaderIds == 0 → 409,不會建立
            - == 1 → 自動 = leaderIds[0]
            - >= 2 → 派工方在 dispatch payload 指定,必須 ∈ leaderIds
        linkedTaskId: { type: string, nullable: true }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/AttachmentRef' }
        submittedAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
        deletedAt: { type: string, format: date-time, nullable: true }
      required: [id, rootOrgId, originatingOrgId, targetOrgId, title, severity, status, requesterId]

    # v1.3 移除:RelayHop schema(隨 relayChain 一併刪除)
    CreateActionRequestRequest:
      type: object
      required: [projectId, title, severity]
      description: |
        自課內提報用(actor 為該 leaf Org 內 user)。
        - 跨層派工請改用 `POST /orgs/{targetOrgId}/dispatch-action-request`(v1.3 single-hop direct dispatch)
        - 自課內提報時 server 自動設 originatingOrgId = targetOrgId = actor 所屬 leaf Org
      properties:
        projectId: { type: string }
        title: { type: string, minLength: 1, maxLength: 200 }
        descriptionMarkdown: { type: string }
        severity:
          type: object
          required: [level]
          properties:
            level: { $ref: '#/components/schemas/SeverityLevel' }
            reason: { type: string }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/AttachmentRef' }
    DispatchActionRequestRequest:
      type: object
      required: [title, severity]
      description: |
        Manager 直派 ActionRequest 到子孫 leaf Org(`targetOrgId` 由 path `/orgs/{orgId}` 帶入,必為 leaf;v1.3)。
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        descriptionMarkdown: { type: string }
        severity:
          type: object
          required: [level]
          properties:
            level: { $ref: '#/components/schemas/SeverityLevel' }
            reason: { type: string }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/AttachmentRef' }
        dueAt: { type: string, format: date-time, nullable: true }
        ownerId:
          type: string
          nullable: true
          description: |
            **條件必填**:當 `targetOrg.leaderIds.length > 1` 時必填,且必須 ∈ targetOrg.leaderIds。
            - leaderIds.length == 1:可省略,server 自動設 = leaderIds[0];若有提供則須等於該值,否則 422
            - leaderIds.length == 0:整個請求 409 `target_org_no_leader`
            - leaderIds.length >= 2:未提供 → 422 `owner_must_be_specified`;不在集合中 → 422 `owner_not_in_leaders`
    UpdateActionRequestRequest:
      type: object
      properties:
        title: { type: string }
        descriptionMarkdown: { type: string }
        severity:
          type: object
          properties:
            level: { $ref: '#/components/schemas/SeverityLevel' }
            reason: { type: string }
        attachments:
          type: array
          items: { $ref: '#/components/schemas/AttachmentRef' }
    ConvertActionRequestToTaskRequest:
      type: object
      required: [taskType, ownerId, assignees]
      properties:
        taskType: { type: string, example: INCIDENT_RESPONSE }
        title: { type: string }
        descriptionMarkdown: { type: string }
        priority: { $ref: '#/components/schemas/Priority' }
        ownerId: { type: string }
        assignees:
          type: array
          minItems: 1
          items: { type: string }
        attributes:
          type: object
          additionalProperties: true
        schedule: { $ref: '#/components/schemas/TimeRange' }
    ActionRequestPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ActionRequest' }
        pageInfo: { $ref: '#/components/schemas/PageInfo' }

    # ---------- Attachment ----------
...